Feature-ID: F-01 Currency conversion from euros to kunas and vice versa

+++ lib/classes/expenses.class.dart
@@ -9,7 +9,7 @@ class Expense
   @HiveField(0)
   String expenseDetails = "";
   @HiveField(1)
-  int expense = 0;
+  double expense = 0;
 
   Expense(this.expenseDetails, this.expense);
 }
